Fixed crt creation function

+++ index.js
@@ -131,9 +131,14 @@ app.listen(3000, function () {
  * Certificate
  */
 
-setTimeout(function() {
-  if (!fs.existsSync(__dirname + '/.http-mitm-proxy/certs/ca.crt')) {
-    var cert = forge.pki.certificateFromPem(fs.readFileSync('.http-mitm-proxy/certs/ca.pem'));
-    fs.writeFile('.http-mitm-proxy/certs/ca.crt', forge.asn1.toDer(forge.pki.certificateToAsn1(cert)).getBytes(), {encoding: 'binary'});
+function createCrt() {
+  if (fs.existsSync(__dirname + '/.http-mitm-proxy/certs/ca.pem')) {
+    if (!fs.existsSync(__dirname + '/.http-mitm-proxy/certs/ca.crt')) {
+      var cert = forge.pki.certificateFromPem(fs.readFileSync('.http-mitm-proxy/certs/ca.pem'));
+      fs.writeFile('.http-mitm-proxy/certs/ca.crt', forge.asn1.toDer(forge.pki.certificateToAsn1(cert)).getBytes(), {encoding: 'binary'});
+    }
+  } else {
+    setTimeout(createCrt, 15000);
   }
-}, 5000);
+}
+createCrt();
